Show a friendlier message on init error
When creating the member extension object in the frontend it needs to
know which section it needs to use as the member info source.

Right now, if the configured value is wrong, you get a weird message
about PHP type checking.

This commit adds a friendlier message when the configuration is not
valid.