Add MembersLogin delegate, for external log in
Usign a section that does not contains a password has been discussed in
issue #274, but the PR did not actually provide a way to let other
extension authenticate the user.

Only if the members section DOES BOT contains a password field, the new
MembersLogin will be called. By default, log in is refused and extension
must explicitly set the required parameters with good values for the
login to succeed.